Fix regression in return type of install_fixtures.
The `FixturesMixin` in `testcase.py` wasn't using the same api as the
rest of the module. I (erik formella) mistakenly changed the return type
of `FixturesManager.install_fixture` thinking it was an internal api so
the test code would work. This broke some people's implementations.

This brings the test Mixin inline with the old api.

tl;dr Charles will hate it but Travis will love it: Adding a test case
to this library caused a regression. </troll>