Add support for lists of fixtures
We already had the ability to define a bunch of fixtures as a list that
would then be accessible as fixture_name_<index>. This adds the
functionality to be able to get the whole list as fixture_name. That way
we can define one-to-many relationships really easily with this list.